use std::set
Use std::set to maintain a sorted list of MemPoolTransactions for
creation of block. The previous implementation was not feasable so
currently using an algorithm that will guarantee the proper order of
transactions but may or may not maximize the profit. It was recommended
to not spend more than 3 days on the problem and thus this shall be the
final change to the algorithm.